push to origin

+++ Comp-programing/HCMUT-ICPC/MissingNumber.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 typedef long long ll;
 int main(){
-    ll n; cin >> n;
+    int n; cin >> n;
     int a[200000] = {0};
     int temp;
     for(int i = 0; i < n-1; i++){
